Added missing Javadoc, pushed implementation to internal package
